# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'CI4 Board API'
  description: "CodeIgniter 4 기반 헤드리스 게시판 REST API\n\n## 인증\n로그인 후 발급된 `access_token`을 `Authorization: Bearer {token}` 헤더로 전달합니다.\n\n## 기본 응답 형식\n```json\n{ \"success\": true,  \"data\": {},   \"message\": \"\" }\n{ \"success\": false, \"data\": null, \"message\": \"에러 메시지\" }\n```\n\n## API 구조\n| 경로 접두사 | 설명 | 인증 |\n|------------|------|------|\n| `/api/v1/*` | 사용자·게시판·쪽지·CMS | JWT (일부 공개) |\n| `/api/admin/v1/*` | 관리자 전용 | Admin JWT (role: superadmin/manager) |"
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: '개발 서버'
paths:
...
  '/api/v1/boards/{bbsId}/articles/{articleIdx}/comments':
    get:
      tags:
        - Comment
      summary: '댓글 목록'
      operationId: 8bcb584806f0c3da37ab6e7eb274771e
      parameters:
        -
          name: bbsId
          in: path
          required: true
          schema:
            type: string
        -
          name: articleIdx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '댓글 목록'
    post:
      tags:
        - Comment
      summary: '댓글 작성'
      operationId: 96e42bcca303b2fce53f96eb14f338f8
      parameters:
        -
          name: bbsId
          in: path
          required: true
          schema:
            type: string
        -
          name: articleIdx
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - comment
              properties:
                comment:
                  type: string
              type: object
      responses:
        '201':
          description: '댓글 작성 완료'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  '/api/v1/boards/{bbsId}/articles/{articleIdx}/comments/{commentIdx}':
    delete:
      tags:
        - Comment
      summary: '댓글 삭제 (작성자만)'
      operationId: b290d7f4203a1e0841e50c6817fdb497
      parameters:
        -
          name: bbsId
          in: path
          required: true
          schema:
            type: string
        -
          name: articleIdx
          in: path
          required: true
          schema:
            type: integer
        -
          name: commentIdx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '삭제 완료'
      security:
        -
          BearerAuth: []
...
  /api/v1/messages/inbox:
    get:
      tags:
        - Message
      summary: '받은 쪽지함'
      operationId: 49feff6fe92b06f653ac89603e9f5e3e
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: '받은 쪽지 목록'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  /api/v1/messages/sent:
    get:
      tags:
        - Message
      summary: '보낸 쪽지함'
      operationId: 83d6c0eef147be5e22ab4419274ec40b
      responses:
        '200':
          description: '보낸 쪽지 목록'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        -
          BearerAuth: []
  '/api/v1/messages/{idx}':
    get:
      tags:
        - Message
      summary: '쪽지 상세 조회 (송·수신자만)'
      operationId: 9f727a028260777d7c9f98c5f1c45143
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '쪽지 상세'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - Message
      summary: '쪽지 삭제 (송·수신자만)'
      operationId: 36ca9d5de43e21dfa6bcc17499096f43
      parameters:
        -
          name: idx
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: '삭제 완료'
      security:
        -
          BearerAuth: []
  /api/v1/messages:
    post:
      tags:
        - Message
      summary: '쪽지 보내기'
      operationId: 1384f41662e3af00bc928c8ed178ed43
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - to
                - contents
              properties:
                to:
                  description: '수신자 user_id 또는 nickname'
                  type: string
                title:
                  type: string
                contents:
                  type: string
              type: object
      responses:
        '201':
          description: '발송 완료'
        '404':
          description: '수신자 없음'
        '422':
          description: '자신에게 발송 불가 또는 필수값 누락'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
    ApiResponse:
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
          nullable: true
      type: object
...
  responses:
    NotFound:
      description: '리소스를 찾을 수 없습니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: '인증이 필요합니다.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      description: 'JWT Bearer token'
      bearerFormat: JWT
      scheme: bearer
...